move bootstrap code to lib/src
