+++ lib/src/bootstrap/browser.dart
@@ -1,9 +1,9 @@
 // Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-import "../src/runner/browser/post_message_channel.dart";
-import "../src/runner/plugin/remote_platform_helpers.dart";
-import "../src/utils.dart";
+import "../runner/browser/post_message_channel.dart";
+import "../runner/plugin/remote_platform_helpers.dart";
+import "../utils.dart";
 
 /// Bootstraps a browser test to communicate with the test runner.
 ///
