+++ lib/src/bootstrap/vm.dart
@@ -5,9 +5,9 @@ import "dart:isolate";
 
 import "package:stream_channel/stream_channel.dart";
 
-import "../src/runner/plugin/remote_platform_helpers.dart";
-import "../src/runner/vm/catch_isolate_errors.dart";
-import "../src/utils.dart";
+import "../runner/plugin/remote_platform_helpers.dart";
+import "../runner/vm/catch_isolate_errors.dart";
+import "../utils.dart";
 
 /// Bootstraps a vm test to communicate with the test runner.
 ///
